reassign randomization code to base Morty class
We had the stat randomization code in the __repr__ function, which ofc only represented it that way. When calling pick_a_winner in morty_war (funktions.py), we were getting the default HP (which was also the class's minimum stat (assuming 0 iv's.)) That was causing a lot of ties and ofc, not the functionality we intended. This commit "bakes in" the random iv's and their stat adjustments to the base Morty class.